backstage: refacotr meta.tsx, remove x-dash templates and keep relevant logic to meta in one file, extend Meta prop type to cover all the prop types and enforce correct ts checking on meta.tsx props
